Create HW1.c
The user interface will ask you to enter ten different character strings using the keyboard. A strlen() value between 2 and 25 inclusive is required for each character string; otherwise, an error notice is shown and the user is questioned again. The program sorts this collection of 10 character strings (based on strcmp()) and then prints all ten strings again in the user-specified ascending OR descending order. The character strings with the lowest and highest ascii values, as determined by strcmp(), are then printed and labeled. The following characters CANNOT be used in character strings: '!', '@', '#', '$', '%', '', '(', or ')'.